credit_approval/requirements.txt
Django>=4.0
djangorestframework
pandas
celery
redis
psycopg2-binary
openpyxl
